requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
requests
rank-bm25
scikit-learn
fastembed
pytest
groq
ragas
pandas
numpy
datasets
